issue-22
Comment: a lot of refinemet was needed, I'm quite hapy now. We just need to solve issue-21 and keep track of success/failures.
Of course we need some tests too. A bit of refactoring in main method of AbstractUnitRunner would be great to make it easier to test of course, but I'll take care during next code session when I'l write tests too.

git-svn-id: https://testedby.googlecode.com/svn/trunk@81 fca47524-2756-0410-b7c8-b385b6c05b64